lmr/requirements.txt
asgiref==3.6.0
Django==3.2.16
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3